Remove omg test file, add .gitignore, benchmark_test.go, bom.go, and bom_test.go
This commit includes the deletion of the omg_test.go file. A new .gitignore file is added to the pkg/tablewriter directory. Additionally, benchmark_test.go, bom.go, and bom_test.go files are created in the pkg/runewidth and pkg/bom directories respectively. This optimizes certain sections of the codebase and introduces necessary tests.